Refactor council prompts, APIs and models to seperate folder
