+++ backend/apis/list_conversations.py
@@ -1,0 +1,16 @@
+"""List conversations endpoint."""
+
+from typing import List
+from fastapi import APIRouter
+
+from .. import storage
+from ..models import ConversationMetadata
+
+router = APIRouter()
+
+
+@router.get("/api/conversations", response_model=List[ConversationMetadata])
+async def list_conversations():
+    """List all conversations (metadata only)."""
+    return storage.list_conversations()
+
