+++ backend/apis/send_message.py
@@ -1,0 +1,54 @@
+"""Send message endpoint."""
+
+from fastapi import APIRouter, HTTPException
+
+from .. import storage
+from ..council import run_full_council, generate_conversation_title
+from ..models import SendMessageRequest
+
+router = APIRouter()
+
+
+@router.post("/api/conversations/{conversation_id}/message")
+async def send_message(conversation_id: str, request: SendMessageRequest):
+    """
+    Send a message and run the 3-stage council process.
+    Returns the complete response with all stages.
+    """
+    # Check if conversation exists
+    conversation = storage.get_conversation(conversation_id)
+    if conversation is None:
+        raise HTTPException(status_code=404, detail="Conversation not found")
+
+    # Check if this is the first message
+    is_first_message = len(conversation["messages"]) == 0
+
+    # Add user message
+    storage.add_user_message(conversation_id, request.content)
+
+    # If this is the first message, generate a title
+    if is_first_message:
+        title = await generate_conversation_title(request.content)
+        storage.update_conversation_title(conversation_id, title)
+
+    # Run the 3-stage council process
+    stage1_results, stage2_results, stage3_result, metadata = await run_full_council(
+        request.content
+    )
+
+    # Add assistant message with all stages
+    storage.add_assistant_message(
+        conversation_id,
+        stage1_results,
+        stage2_results,
+        stage3_result
+    )
+
+    # Return the complete response with metadata
+    return {
+        "stage1": stage1_results,
+        "stage2": stage2_results,
+        "stage3": stage3_result,
+        "metadata": metadata
+    }
+
