+++ backend/models/__init__.py
@@ -1,0 +1,14 @@
+"""Pydantic models for LLM Council API."""
+
+from .create_conversation_request import CreateConversationRequest
+from .send_message_request import SendMessageRequest
+from .conversation_metadata import ConversationMetadata
+from .conversation import Conversation
+
+__all__ = [
+    "CreateConversationRequest",
+    "SendMessageRequest",
+    "ConversationMetadata",
+    "Conversation",
+]
+
