+++ backend/models/conversation_metadata.py
@@ -1,0 +1,12 @@
+"""Model for conversation metadata in list view."""
+
+from pydantic import BaseModel
+
+
+class ConversationMetadata(BaseModel):
+    """Conversation metadata for list view."""
+    id: str
+    created_at: str
+    title: str
+    message_count: int
+
