+++ backend/models/create_conversation_request.py
@@ -1,0 +1,9 @@
+"""Request model for creating a new conversation."""
+
+from pydantic import BaseModel
+
+
+class CreateConversationRequest(BaseModel):
+    """Request to create a new conversation."""
+    pass
+
