+++ backend/models/send_message_request.py
@@ -1,0 +1,9 @@
+"""Request model for sending a message in a conversation."""
+
+from pydantic import BaseModel
+
+
+class SendMessageRequest(BaseModel):
+    """Request to send a message in a conversation."""
+    content: str
+
