+++ backend/prompts/__init__.py
@@ -1,0 +1,1 @@
+"""Prompt templates loaded from markdown files."""
