+++ backend/prompts/title_generation.md
@@ -1,0 +1,7 @@
+Generate a very short title (3-5 words maximum) that summarizes the following question.
+The title should be concise and descriptive. Do not use quotes or punctuation in the title.
+
+Question: {user_query}
+
+Title:
+
